Fix concurrency issue in Jdk8Cleaner
Motivation
----------
The Jdk8Cleaner.references Set was accessed
from multiple threads in an unsafe way.

Modifications
-------------
Change the backing map from an IdentityHashMap to
a ConcurrentHashMap. Don't need the "identity" behavior
since the keys don't implement equals or hashcode.

Document the purpose of the `references` Set, since at
first glance it appears to be useless. (Fortunately,
the unit test fails if it's removed.)

Change-Id: If2af83de583ec361364a922c81fad78473798c42
Reviewed-on: http://review.couchbase.org/c/java-couchbase-encryption/+/143917
Tested-by: David Nault <david.nault@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>