Make core-io a "provided" dependency
Motivation
----------
Don't want Maven users to get into trouble by
declaring the `couchbase-encryption` and `java-client`
dependencies in the wrong order and ending up with
a version of `core-io` that's not compatible with
the `java-client` version.

Modifications
-------------
Mark `core-io` as `provided`. It will most likely
be present, since the user will also be depending on
`java-client` (or some other jvm client) which has
`core-io` as a transitive dependency.

Build against the earliest version of core-io we
support: 2.0.6, which corresponds to SDK 3.0.5.

Change-Id: I15d0f9043747d6a89bc3a789916814fc51f6fc13
Reviewed-on: http://review.couchbase.org/c/java-couchbase-encryption/+/143925
Tested-by: David Nault <david.nault@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>