Deprecate Keyring.rotating in favor of RotatingKeyring
Motivation
----------
The Keyring.rotating() decorator has a flaw that makes it
unsuitable for use in distributed deployments. It always
encrypts using the latest version of a key, regardless of
whether that version is available yet on other nodes.

Instead, encourage users to use FilesystemRotatingKeyring
or extend RotatingKeyring. These classes let the user designate
one version of the key as "primary". The primary version is used
to encrypt data, even if it's not the latest. When a new key version
is fully deployed across all nodes, only then should the user
designate that version as "primary".

These classes originated in the couchbaselabs/java-rabbit-rolodex project.

Change-Id: I658509e22ed4ff77227cf1a66f97afe16fdfa307
Reviewed-on: https://review.couchbase.org/c/java-couchbase-encryption/+/175129
Tested-by: David Nault <david.nault@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>